Initial add of kotlin application based on getting started guide.
https://www.dropwizard.io/en/latest/getting-started.html
Note: It is necesary to call "registerKotlinModule()" on the ObjectMapper.

Added basic config.yml so that its possible to start the application.
I have not added the resource objects for now - since we need to wire up Guicey next.

At this point you can start ExampleApplication in your IDE with the command line:
  server config.yml