Updated pom.xml to resolve issues.
Added filter condition to shade plugin to exclude regular files that appear in multiple jar files.
Added exclusions so that only one version of each class file is included in the output.
Set the main class of the generated application jar file.
Moved the dependency-reduced-pom.xml into the target directory.